refactor: Update GenerateUniqueShortID function to use MongoDB for checking uniqueness
The GenerateUniqueShortID function in the utils package has been updated to use MongoDB for checking the uniqueness of the generated short ID. The function now takes a mongo.Collection as a parameter and uses it to query the database for existing short URLs. This change improves the efficiency and reliability of the ID generation process.